Add the ability to refer to a function while defining it.
This enables recursion as in:
(defun 'f n
  (list
    (print n)
    (f (+ n 1))
  )
)

This is handled by a MaybeFunctionCall class.
When executed, this class will look up its
name in the current global scope and call the actual
function.

This means that the user function doesn't need to be
defined for the call to it to parse.

This allows me to simplify the FizzBuzz example.
Along with "last" so we only see one of the return values.
You can think of the usual body of a function as:
(defun 'f (last (list
  (statement)
  (return value)
)))

Note: the limit of the FizzBuzz example had to be
lowered to 98 because after that we hit Python's
recursion limit.