Minor cleanup, add "not" builtin.
New builtin not:
(not (eq 1 0) => True

New library function neq
(neq 1 2) => True

This has exposed an issue with * args.
If you try to call something with a minimum
number of args >=2 you can't just do:
(foo **)
Since when the call gets built it sees "**"
not as an expansion of a list, but just one argument.

Also realised that a lot of my list functions fall
over on empty/single item lists. Some of them this is
probably correct e.g. head. Others, e.g. reverse, I'm
not so sure.