Add map function.
Along the way fix a bug where when you called
something in a let it was called twice.
Once when it was prepared and again when
it was executed.

This was because let didn't return a new set of
args from prepare.
(let 'f (foo 1 2))

Would call "foo" to get the value to assign to f.
So f now has the correct value, but it returns the
call object still. So when execute continues it calls
foo *again*.

I only noticed this adding prints into map.