Switch to an iterative/fake recursion style for execution.
This means that anything that would be evaluated
depth first doesn't increase the call depth from
Python's point of view. (+ 1 (+ 1)) is just one
call to execute.

What this doesn't handle is any "breadth first"
calls like cond/if, or calling a user function.
Those trigger a new execute call.

It does get us to 136 in FizzBuzz, which has
become the unofficial benchmark it seems.