Execute if/let/cond/defun in one pass.
Calls to builtins don't cause more calls
to execute. Now "breadth" functions don't either.
(calls to user functions do, but the defun for
them does not)

The idea is we keep the same fake stack
loop over the arguments but we sort the arguments
so that those needed for the prepare step come first.

For example, a condition is written as:
(cond (condition 1) (action 1) (condition 2) (action 2))
It will sort its args before execution so that they are:
(condition 1) (condition 2) (action 1) (action 2)

When we know we've executed both conditions, we can
"prepare". Which in cond's case is choosing one (or none,
if no true conditions) of the actions.

Then the normal loop over the arguments continues, with
the chosen condition being the next argument to be
executed as expected.