Function calls without recursion.
MaybeFunctionCalls now have the name/expression
of the function to be called as arg[0].

Then when this and all the args to be passed
have been executed, we lookup the actual function.

Assuming that was succesful append that function
to the end of the MaybeFunctionCall's arg list and
return to execute.

The execute loop then sees this Call object and
handles it as normal.

This means that MaybeFunctionCall is not calling
execute in it's prepare function.
(though it has a pile of hacks in it to impelement
this)

With this change, the only place calling execute is
run_source, which is what we want. See tests.py
"execute_count" for the numbers of execute calls we do.

Long story short, recursion is no longer going to call
execute over and over.