Add Morse Code example.
New library functions:
* find, returns the index of the
  first occurrence of a value in
  a list.
* str_to_list, converts a string
  to a list of chars.

Added more methods to StringVar.
To allow comparison, summing and
indexing. Note that I'm not allowing
slicing (with a runtime check).

Also added a special case to flatten
so that it won't treat StringVar as
iterable. This plus the runtime check
in __iter__ should prevent any surprises
for the time being.

Later I can allow slicing strings
if needed.

Also realised that if the name of a
variable is a StringVar, we must
get its value. The *value* of a var
can stay as a StringVar.
(it was the other way around before)