Add some tests for StringVar handling.
Realised along the way that (flatten <str>)
gives you a list of chars.

Unintended and inconsistent? Sure, but
I'll use it for now and remove the
str_to_list function from the lib.

So the rule is flattening a list that
has literals in it, doesn't change the
literals. Flattening a literal directly,
does.