Got all the ABC block examples right.
Switched to tracking indexes since if you
track values you will "use" more than one
block if two blocks are the same value.

Had to abuse the fact that findif uses a var
"idx" in the scope the predicate executes in.
There should be a better way to do that.

Something like enumerate/zip in Python.