Start an empty scope when calling a function.
Various changes to examples and lib to fix
places where I was relying on outer vars.

The findif predicate now gets the index.
May change this but it's making the ABCBlock
example work for now.

This has shown the need for some kind of
capture semantics for lambda usage.

Hack to get around that for now is to do:
(defun 'get_var <var>)
Since <var> isn't a function call it'll
get resolved at define time and give you
a global getter.

Which I didn't even know worked.