Add lambda functions with capture lists.
A lambda is basically an anonymous function with
a list of variables to capture in place of its name.

(lambda (list 'a 'b) 'c (+ a b c))

The capture list is required and must be a list,
but can be empty. Names are escaped with ' like
the usual var names.

Capturing happens at the time the function is defined.
(let 'x 1
  (let 'fn (lambda (list x) (+ x))
    (let 'x 2
      (fn)
    )
  )
)

Will result in 1, because x was captured at the "let 'fn"
call.

This has allowed me to get rid of a few hacks in findif
and the ABC/Morse examples.